+++ src/main/java/com/ssgsakk/ssgdotcom/cart/vo/CartRequestVo.java
@@ -5,13 +5,12 @@ import lombok.Builder;
 import lombok.Getter;
 
 
-
 @Getter
 @Builder
 public class CartRequestVo {
     private Long optionAndStockSeq;
     private Integer quantity;
     private Long productSeq;
-    private Short checkbox;
-    private Short fixItem;
+    private Integer checkbox;
+    private Integer fixItem;
 }
